airflow/requirements.txt
apache-airflow-providers-google
apache-airflow[google]
pyarrow
nba-api
apache-airflow-providers-postgres
pyspark==3.2.0
pandas